api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
pytesseract==0.3.10
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.3
langsmith==0.0.83
chromadb==0.4.22
sentence-transformers==2.3.1
